+++ src/BubbleScript.js
@@ -19,35 +19,8 @@ const Vector = require("./o/vector");
 const §ymbol = require("./o/symbol");
 const Keyword = require("./o/keyword");
 const Bubble = require("./o/bubble");
-
-const createBinding = require("./f/create_binding.js");
-
-class Macro {
-  constructor(binding, params, body, opts={}) {
-    this.binding = binding;
-    this.params = params.toList();
-    this.body = body;
-    this.name = opts.name;
-  }
-
-  expand(params) {
-    let binding = createBinding(this.binding,
-      this.params, params);
-
-    return this.body.mapEval(binding);
-  }
-
-  toString() {
-    return "(macro " + this.args.toString() +
-      this.body.toString() + ")";
-  }
-}
-
-class MacroExpanded {
-  constructor(expanded) {
-    this.expanded = expanded;
-  }
-}
+const Fn = require("./o/fn");
+const Macro = require("./o/macro");
 
 
 function tokenize(inputString) {
